@@ -12,17 +12,17 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "9.1.13",
-    "@angular/common": "^9.1.13",
-    "@angular/compiler": "9.1.13",
-    "@angular/core": "9.1.13",
-    "@angular/forms": "9.1.13",
-    "@angular/platform-browser": "9.1.13",
-    "@angular/platform-browser-dynamic": "9.1.13",
-    "@angular/platform-server": "9.1.13",
-    "@angular/router": "9.1.13",
+    "@angular/animations": "13.0.3",
+    "@angular/common": "^13.0.3",
+    "@angular/compiler": "13.0.3",
+    "@angular/core": "13.0.3",
+    "@angular/forms": "13.0.3",
+    "@angular/platform-browser": "13.0.3",
+    "@angular/platform-browser-dynamic": "13.0.3",
+    "@angular/platform-server": "13.0.3",
+    "@angular/router": "^13.0.3",
     "@nguniversal/module-map-ngfactory-loader": "8.1.1",
-    "@ngx-translate/core": "^14.0.0",
+    "@ngx-translate/core": "^8.0.0",
     "@ngx-translate/http-loader": "^7.0.0",
     "bootstrap": "^4.6.0",
     "core-js": "^3.8.3",
@@ -32,34 +32,35 @@
     "oidc-client": "^1.11.3",
     "popper.js": "^1.16.0",
     "rxjs": "^6.6.7",
-    "tslib": "^1.10.0",
-    "zone.js": "~0.10.2"
+    "rxjs-compat": "^6.6.7",
+    "tslib": "^2.0.0",
+    "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.901.15",
-    "@angular/cli": "^9.1.15",
-    "@angular/compiler-cli": "^9.1.13",
-    "@angular/language-service": "^9.1.13",
+    "@angular-devkit/build-angular": "^13.0.4",
+    "@angular/cli": "^13.0.4",
+    "@angular/compiler-cli": "^13.0.3",
+    "@angular/language-service": "^13.0.3",
     "@biesbjerg/ngx-translate-extract": "^7.0.4",
-    "@types/jasmine": "~3.4.4",
+    "@types/jasmine": "~3.6.0",
     "@types/jasminewd2": "~2.0.8",
     "@types/node": "^12.11.1",
-    "codelyzer": "^5.1.2",
+    "codelyzer": "^6.0.0",
     "ini": "^1.3.7",
-    "jasmine-core": "~3.5.0",
-    "jasmine-spec-reporter": "~4.2.1",
-    "karma": "^5.2.3",
+    "jasmine-core": "~3.6.0",
+    "jasmine-spec-reporter": "~5.0.0",
+    "karma": "~6.3.9",
     "karma-chrome-launcher": "~3.1.0",
-    "karma-coverage-istanbul-reporter": "~2.1.0",
-    "karma-jasmine": "~2.0.1",
-    "karma-jasmine-html-reporter": "^1.5.4",
-    "typescript": "3.8.3"
+    "karma-coverage-istanbul-reporter": "~3.0.2",
+    "karma-jasmine": "~4.0.0",
+    "karma-jasmine-html-reporter": "^1.5.0",
+    "typescript": "4.4.4"
   },
   "optionalDependencies": {
-    "node-sass": "^5.0.0",
-    "protractor": "~5.4.2",
+    "node-sass": "^7.0.0",
+    "protractor": "~7.0.0",
     "ts-node": "~8.4.1",
-    "tslint": "~5.20.0"
+    "tslint": "~6.1.0"
   },
   "resolutions": {
     "url-parse": ">=1.5.0",
